project/api/views.py
from rest_framework import viewsets
from api.models import Result
from api.serializers import ResultSerializer


class ResultViewSet(viewsets.ModelViewSet):
    queryset = Result.objects.all()
    serializer_class = ResultSerializer