requirements.txt
Django==2.1
django-webpack-loader==0.6.0
djangorestframework==3.8.2
